Implement new node API
The basic interface for the API is two channels, one that sends abstract
requests to the node and one that receives notifications.

On top of this abstract API we can implement both the existing
bi-directional RPC api, or a uni-directional one where a client opens 2
connections to the node, one for submitting requests and one for
receiving notifications. A sample HTTP implementation for the latter
using chunked transfer encoding is provided.